requirements.txt
# Core dependencies
pyautogen>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0

# Web interface
streamlit>=1.28.0

# Static analysis tools
pylint>=3.0.0
pycodestyle>=2.11.0
flake8>=6.1.0
bandit>=1.7.5
radon>=6.0.1

# Testing tools
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-json-report>=1.5.0

# Code execution (optional - for sandbox)
docker>=7.0.0

# Utilities
typing-extensions>=4.8.0
pydantic>=2.0.0

# Development dependencies (optional)
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0